refactor: enable deprecation for old Java interop
