+++ main/src/ca/uwaterloo/flix/language/errors/NameError.scala
@@ -30,6 +30,25 @@ sealed trait NameError extends CompilationMessage with Recoverable {
 object NameError {
 
   /**
+    * An error raised to indicate a deprecated feature
+    * *
+    * @param loc the location of the deprecated feature.
+    */
+  case class Deprecated(loc: SourceLocation) extends NameError with Recoverable {
+    def summary: String = s"Deprecated feature."
+
+    def message(formatter: Formatter): String = {
+      import formatter._
+      s""">> Deprecated feature. Use --Xdeprecated to enable.
+         |
+         |${code(loc, "deprecated")}
+         |""".stripMargin
+    }
+
+    override def explain(formatter: Formatter): Option[String] = None
+  }
+
+  /**
     * An error raised to indicate that the given `name` is defined multiple time.
     *
     * @param name the name.
