+++ main/src/ca/uwaterloo/flix/language/errors/WeederError.scala
@@ -29,26 +29,6 @@ sealed trait WeederError extends CompilationMessage {
 
 object WeederError {
 
-
-  /**
-   * An error raised to indicate a deprecated feature
-   * *
-   * @param loc the location of the deprecated feature.
-   */
-  case class Deprecated(loc: SourceLocation) extends WeederError with Recoverable {
-    def summary: String = s"Deprecated feature."
-
-    def message(formatter: Formatter): String = {
-      import formatter._
-      s""">> Deprecated feature. Use --Xdeprecated to enable.
-         |
-         |${code(loc, "deprecated")}
-         |""".stripMargin
-    }
-
-    override def explain(formatter: Formatter): Option[String] = None
-  }
-
   /**
     * An error raised to indicate that the annotation `name` was used multiple times.
     *
